// apps/frontend/src/shared/types/event-sessions/chat-event-payload.type.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UsageInfo } from "../ai-client/usage-info.type";

export type ChatEventPayload = { "event": "chunk", chat_id: string, branch_id: string, message_id: string, message_content: string, } | { "event": "usage", chat_id: string, branch_id: string, usage_data: UsageInfo, } | { "event": "done", chat_id: string, branch_id: string, };