// apps/frontend/src/shared/types/nodes/node.type.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeRole } from "./node-role.type";
import type { NodeType } from "./node-type.type";

export type Node = { id: string, type: NodeType, role: NodeRole, branchId: string, messageId: string | null, messageContent: string | null, messageVersionNumber: bigint | null, messageVersionCount: bigint | null, extensions: string | null, updatedAt: string, };